♻️ Refactor: 관심사 비슷한 훅 병합
refactor: 관심사 비슷한 훅 병합
