+++ src/pages/QuestionForm/hooks/useQuestionReorder.ts
@@ -1,19 +1,47 @@
-import { useCallback, useMemo } from "react";
+import { useCallback, useEffect, useMemo, useRef, useState } from "react";
 import type { Question } from "../../../types/survey";
 
 interface UseQuestionReorderProps {
 	questions: Question[];
 	onReorder: (questions: Question[]) => void;
+	delay?: number;
 }
 
 export const useQuestionReorder = ({
 	questions,
 	onReorder,
+	delay = 600,
 }: UseQuestionReorderProps) => {
 	const sortedQuestions = useMemo(
 		() => [...questions].sort((a, b) => a.questionOrder - b.questionOrder),
 		[questions],
 	);
+
+	const [displayOrderMap, setDisplayOrderMap] = useState<
+		Record<number, number>
+	>({});
+	const hasInitialized = useRef(false);
+
+	useEffect(() => {
+		const newOrderMap: Record<number, number> = {};
+		sortedQuestions.forEach((question, index) => {
+			newOrderMap[question.questionId] = index;
+		});
+
+		if (!hasInitialized.current) {
+			setDisplayOrderMap(newOrderMap);
+			hasInitialized.current = true;
+			return;
+		}
+
+		const timer = window.setTimeout(() => {
+			setDisplayOrderMap(newOrderMap);
+		}, delay);
+
+		return () => {
+			window.clearTimeout(timer);
+		};
+	}, [sortedQuestions, delay]);
 
 	const handleMoveUp = useCallback(
 		(e: React.MouseEvent, currentIndex: number) => {
@@ -57,6 +85,7 @@ export const useQuestionReorder = ({
 
 	return {
 		sortedQuestions,
+		displayOrderMap,
 		handleMoveUp,
 		handleMoveDown,
 	};
